move package bg to worker/internal/eventlogs
