+++ cmd/worker/internal/eventlogs/BUILD.bazel
@@ -1,18 +1,36 @@
+load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "eventlogs",
-    srcs = ["job.go"],
+    srcs = [
+        "janitor.go",
+        "job.go",
+    ],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/worker/internal/eventlogs",
     visibility = ["//cmd/worker:__subpackages__"],
     deps = [
         "//cmd/worker/job",
         "//cmd/worker/shared/init/db",
-        "//internal/bg",
+        "//internal/conf",
         "//internal/database",
         "//internal/env",
         "//internal/goroutine",
         "//internal/metrics",
         "//internal/observation",
+    ],
+)
+
+go_test(
+    name = "eventlogs_test",
+    srcs = ["janitor_test.go"],
+    embed = [":eventlogs"],
+    deps = [
+        "//internal/conf",
+        "//internal/database",
+        "//internal/database/dbtest",
+        "//schema",
+        "@com_github_sourcegraph_log//logtest",
+        "@com_github_stretchr_testify//require",
     ],
 )
