+++ cmd/worker/internal/eventlogs/janitor.go
@@ -1,4 +1,4 @@
-package bg
+package eventlogs
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 )
 
-func DeleteOldEventLogsInPostgres(ctx context.Context, db database.DB) error {
+func deleteOldEventLogsInPostgres(ctx context.Context, db database.DB) error {
 	// We choose 93 days as the interval to ensure that we have at least the last three months
 	// of logs at all times.
 	_, err := db.ExecContext(
@@ -17,7 +17,7 @@ func DeleteOldEventLogsInPostgres(ctx context.Context, db database.DB) error {
 	return err
 }
 
-func DeleteOldSecurityEventLogsInPostgres(ctx context.Context, db database.DB) error {
+func deleteOldSecurityEventLogsInPostgres(ctx context.Context, db database.DB) error {
 	// Only clean up if security event logs are being stored in the database.
 	c := conf.Get()
 	if c.Log == nil || c.Log.SecurityEventLog == nil {
