--- internal/bg/BUILD.bazel
@@ -1,28 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-load("//dev:go_defs.bzl", "go_test")
-
-go_library(
-    name = "bg",
-    srcs = ["delete_old_event_logs_in_postgres.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/bg",
-    visibility = ["//:__subpackages__"],
-    deps = [
-        "//internal/conf",
-        "//internal/database",
-    ],
-)
-
-go_test(
-    name = "bg_test",
-    srcs = ["delete_old_event_logs_in_postgres_test.go"],
-    embed = [":bg"],
-    tags = ["requires-network"],
-    deps = [
-        "//internal/conf",
-        "//internal/database",
-        "//internal/database/dbtest",
-        "//schema",
-        "@com_github_sourcegraph_log//logtest",
-        "@com_github_stretchr_testify//require",
-    ],
-)
